ADDSLOTS/DELSLOTS* returns error when myself is not a master
ADDSLOTS / DELSLOTS / ADDSLOTSRANGE / DELSLOTSRANGE commands
should be used in primary. In replica, the final slot ownership
will be gossip from the primary, there commands should not be
used in replicas.

This just adds another layer of validation so that the users
get the error sooner so they don't send the wrong node.